chang user_id to long

+++ src/main/java/com/wequan/bu/im/db/Message.java
@@ -10,23 +10,23 @@ import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBTable;
  */
 @DynamoDBTable(tableName = "msg_sync")
 public class Message {
-    private Integer user_id;
-    private Integer msg_id;
+    private Long user_id;
+    private Long msg_id;
     private String data_content;
 
-    public Message(Integer user_id, Integer msg_id, String data_content) {
+    public Message(Long user_id, Long msg_id, String data_content) {
         this.user_id = user_id;
         this.msg_id = msg_id;
         this.data_content = data_content;
     }
 
     @DynamoDBHashKey(attributeName = "user_id")
-    public Integer getUserId() {
+    public Long getUserId() {
         return this.user_id;
     }
 
     @DynamoDBRangeKey(attributeName = "msg_id")
-    public Integer getMsgId() {
+    public Long getMsgId() {
         return this.msg_id;
     }
